refactor(#800): split providers/cursor.rs by moving tests to sibling
Move 3,128-line providers/cursor.rs into providers/cursor/ directory:

- cursor/mod.rs:   2,286 LOC (production code)
- cursor/tests.rs:   841 LOC (unit tests)

Both pieces well below the 2,500 LOC limit. No behavior changes.

`cargo test -p budi-core --lib` green (737 pass; the one pre-existing
flake — providers::cursor::tests::ingest_roundtrip_writes_embedded_or_manifest_source
— reproduces on `main` too and is unrelated to this split).
